Fix v7.1.6 native discharge confirmation handling
