+++ .github/scripts/test_issue530_transport_sweep_discharge_confirmation_runtime.js
@@ -1,0 +1,18 @@
+#!/usr/bin/env node
+'use strict';
+const assert=require('node:assert/strict');const fs=require('node:fs');const path=require('node:path');const vm=require('node:vm');
+const root=path.resolve(__dirname,'..','..');const source=fs.readFileSync(path.join(root,'src','MissionChief_Map_Command_Toolkit.user.js'),'utf8');
+function extractFunction(name){const marker=`    function ${name}(`;const start=source.indexOf(marker);assert.ok(start>=0,`${name} missing`);const signatureEnd=source.indexOf(') {',start);assert.ok(signatureEnd>=0);const open=signatureEnd+2;let depth=0,quote='',escaped=false;for(let i=open;i<source.length;i++){const c=source[i];if(quote){if(escaped)escaped=false;else if(c==='\\')escaped=true;else if(c===quote)quote='';continue;}if(c==='"'||c==="'"||c==='`'){quote=c;continue;}if(c==='{')depth++;if(c==='}'&&--depth===0)return source.slice(start,i+1);}throw new Error(`Could not extract ${name}`);}
+function control(label){return{textContent:label,value:'',title:'',isConnected:true,disabled:false,clicks:0,attrs:{},click(){this.clicks+=1;},getAttribute(name){return this.attrs[name]||null;}};}
+function dialog(options={}){const abort=control('Abort');const disable=control('discharge and disable confirmation');const confirm=control('Yes, discharge!');const controls=options.controls||[abort,disable,confirm];return{isConnected:true,textContent:options.text||'Discharge patient Do you really want to discharge the patient of this vehicle? No payment will be credited for the patient! Abort discharge and disable confirmation Yes, discharge!',controls,closest(){return null;},querySelectorAll(){return controls;},abort,disable,confirm};}
+let roots=[];const logs=[];const runtime={running:true,pendingDischargeKey:'101:201',confirmedDischargeDialogKeys:new Set()};
+const sandbox={console,String,Array,Map,Set,SCRIPT:{panelId:'toolkit-panel'},transportSweepRuntime:runtime,normaliseTransportSweepReleaseText:value=>String(value||'').replace(/\s+/g,' ').trim().toLowerCase(),transportSweepElementVisible:element=>Boolean(element?.isConnected),transportSweepVisibleWindowRoots:()=>roots,transportSweepDocumentContexts:()=>[],transportSweepLog:message=>logs.push(message)};
+const constants=`const TRANSPORT_SWEEP_DISCHARGE_DIALOG_TITLE='discharge patient';const TRANSPORT_SWEEP_DISCHARGE_DIALOG_WARNING='do you really want to discharge the patient of this vehicle? no payment will be credited for the patient!';const TRANSPORT_SWEEP_DISCHARGE_DIALOG_ABORT='abort';const TRANSPORT_SWEEP_DISCHARGE_DIALOG_DISABLE='discharge and disable confirmation';const TRANSPORT_SWEEP_DISCHARGE_DIALOG_CONFIRM='yes, discharge!';`;
+vm.createContext(sandbox);vm.runInContext(`${constants}\n${extractFunction('transportSweepDialogControlText')}\n${extractFunction('transportSweepDischargeConfirmationRoots')}\n${extractFunction('clickTransportSweepDischargeConfirmation')}\nthis.clickTransportSweepDischargeConfirmation=clickTransportSweepDischargeConfirmation;`,sandbox);const click=sandbox.clickTransportSweepDischargeConfirmation;
+const exact=dialog();roots=[exact];assert.equal(click('101:201'),true);assert.equal(exact.confirm.clicks,1);assert.equal(exact.abort.clicks,0);assert.equal(exact.disable.clicks,0);assert.equal(logs.length,1);
+assert.equal(click('101:201'),false);assert.equal(exact.confirm.clicks,1,'Repeated scans must not double-click');
+runtime.pendingDischargeKey='101:202';const unrelated=dialog({text:'Delete building Abort discharge and disable confirmation Yes, discharge!'});roots=[unrelated];assert.equal(click('101:202'),false);assert.equal(unrelated.confirm.clicks,0,'Unrelated modal must be ignored');
+roots=[];assert.equal(click('101:202'),false,'No dialog is a valid no-op');const delayed=dialog();roots=[delayed];assert.equal(click('101:202'),true,'A later dialog scan must confirm');assert.equal(delayed.confirm.clicks,1);
+runtime.running=false;runtime.pendingDischargeKey='101:203';const inactive=dialog();roots=[inactive];assert.equal(click('101:203'),false);assert.equal(inactive.confirm.clicks,0,'Inactive sweep must not confirm');
+runtime.running=true;runtime.pendingDischargeKey='different';assert.equal(click('101:203'),false);assert.equal(inactive.confirm.clicks,0,'Unarmed release key must not confirm');
+console.log('Issue #530 discharge confirmation runtime contract passed.');
