--- .github/workflows/apply-issue530-discharge-confirmation.yml
@@ -1,73 +1,0 @@
-name: Apply Issue 530 Discharge Confirmation Fix
-
-on:
-  pull_request:
-    branches: [main]
-    paths:
-      - ".github/issue530/apply_fix.py"
-      - ".github/workflows/apply-issue530-discharge-confirmation.yml"
-
-permissions:
-  contents: read
-
-concurrency:
-  group: toolkit-issue-530-discharge-confirmation
-  cancel-in-progress: false
-
-jobs:
-  apply:
-    if: github.actor == 'Conroy1988' && github.head_ref == 'fix/issue-530-transport-sweep-discharge-confirmation'
-    runs-on: ubuntu-latest
-    timeout-minutes: 30
-    steps:
-      - name: Check out exact Issue 530 branch
-        uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7
-        with:
-          ref: fix/issue-530-transport-sweep-discharge-confirmation
-          fetch-depth: 0
-          persist-credentials: false
-
-      - name: Apply scoped native discharge confirmation handling
-        run: python3 .github/issue530/apply_fix.py
-
-      - name: Validate complete v7.1.6 candidate
-        shell: bash
-        run: |
-          set -euo pipefail
-          python3 .github/scripts/validate_userscript.py
-          bash .github/scripts/run_userscript_preflight.sh --contracts
-          python3 .github/scripts/test_documentation_consistency.py
-          node --check src/MissionChief_Map_Command_Toolkit.user.js
-          cmp --silent dist/MissionChief_Map_Command_Toolkit.user.js dist/MissionChief_Map_Command_Toolkit.txt
-
-      - name: Enforce unchanged performance envelope
-        shell: bash
-        run: |
-          set -euo pipefail
-          git fetch origin main --depth=1
-          git show origin/main:src/MissionChief_Map_Command_Toolkit.user.js > /tmp/performance-base.user.js
-          python3 .github/scripts/test_performance_budget.py
-          python3 .github/scripts/check_performance_budget.py \
-            --candidate src/MissionChief_Map_Command_Toolkit.user.js \
-            --base /tmp/performance-base.user.js \
-            --policy .github/performance-budget.json \
-            --json-output /tmp/issue530-performance.json \
-            --markdown-output /tmp/issue530-performance.md
-          cat /tmp/issue530-performance.md
-
-      - name: Commit exact v7.1.6 candidate
-        env:
-          DEVELOPMENT_PR_TOKEN: ${{ secrets.DEVELOPMENT_PR_TOKEN }}
-        shell: bash
-        run: |
-          set -euo pipefail
-          test -n "$DEVELOPMENT_PR_TOKEN"
-          test "$GITHUB_HEAD_REF" = "fix/issue-530-transport-sweep-discharge-confirmation"
-          git config user.name "Conroy1988"
-          git config user.email "27301455+Conroy1988@users.noreply.github.com"
-          git add -A
-          git diff --cached --quiet && { echo "No Issue 530 changes produced."; exit 1; }
-          git commit -m "Fix v7.1.6 native discharge confirmation handling"
-          BASIC_AUTH="$(printf 'x-access-token:%s' "$DEVELOPMENT_PR_TOKEN" | base64 -w0)"
-          git -c http.https://github.com/.extraheader="AUTHORIZATION: basic ${BASIC_AUTH}" \
-            push origin HEAD:refs/heads/fix/issue-530-transport-sweep-discharge-confirmation
